Add files via upload
@caitobrien and @JLGnotes 
Here is a block of code that pulls SNOTEL data , determines the day-of-year in each year when snowpack (as measured by SWE) is greatest.
These results are written to a csv file.
It must be run completely once for all historical years to be cached. 
An update can be done annual by limiting the years downloaded to the missing year. It can be run after mid-May for an annual update.